Fix style for commands
