+++ src/AudioBand/GlobalSuppressions.cs
@@ -7,4 +7,8 @@
 [assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1600:Elements must be documented", Justification = "None", Scope = "member", Target = "~M:AudioBand.MainControl.#ctor")]
 [assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1600:Elements must be documented", Justification = "None", Scope = "member", Target = "~M:AudioBand.MainControl.OnResize(System.EventArgs)")]
 [assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.OrderingRules", "SA1205:Partial elements must declare access", Justification = "None", Scope = "type", Target = "~T:AudioBand.MainControl")]
+[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.LayoutRules", "SA1502:Element must not be on a single line", Justification = "None", Scope = "member", Target = "~M:AudioBand.Commands.AsyncRelayCommand`1.#ctor(System.Func{`0,System.Threading.Tasks.Task})")]
+[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.LayoutRules", "SA1502:Element must not be on a single line", Justification = "None", Scope = "member", Target = "~M:AudioBand.Commands.RelayCommand`1.#ctor(System.Action{`0})")]
+[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.LayoutRules", "SA1502:Element must not be on a single line", Justification = "None", Scope = "member", Target = "~M:AudioBand.Commands.RelayCommand.#ctor(System.Action{System.Object})")]
+[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.LayoutRules", "SA1502:Element must not be on a single line", Justification = "None", Scope = "member", Target = "~M:AudioBand.Commands.RelayCommand.#ctor(System.Action{System.Object},System.Predicate{System.Object})")]
